Refactor common codes
